refactor: improve type safety and fix template literal errors
