+++ lib/schemas/search.ts
@@ -1,14 +1,11 @@
-import { z } from "zod";
+/**
+ * Search Schema Utilities
+ *
+ * IMPORTANT: All Zod schemas have been moved to types/ directory for centralized type system.
+ * This file now only contains utility functions that work with the schemas.
+ *
+ * @see types/search.ts for actual Zod schemas
+ */
 
-export const searchResultItemSchema = z.object({
-  id: z.string(),
-  type: z.enum(["bookmark", "blog-post", "project", "page"]),
-  title: z.string(),
-  description: z.string().optional(),
-  url: z.string(),
-  score: z.number(),
-  highlights: z.array(z.string()).optional(),
-  metadata: z.record(z.unknown()).optional(),
-});
-
-export const searchResultsSchema = z.array(searchResultItemSchema);
+// This file intentionally left minimal - all schemas moved to types/search.ts
+// to comply with centralized type system requirements
